+++ lib/domain/services/date_extraction/date_extractor_service.dart
@@ -1,8 +1,8 @@
 import 'dart:io';
 
-export 'exif_extractor.dart';
-export 'guess_extractor.dart';
-export 'json_extractor.dart';
+export 'exif_date_extractor.dart';
+export 'filename_date_extractor.dart';
+export 'json_date_extractor.dart';
 
 /// Function that can take a file and potentially extract DateTime of it
 typedef DateTimeExtractor = Future<DateTime?> Function(File);