+++ test/remove_extra_regex_test.dart
@@ -1,6 +1,6 @@
 import 'dart:convert';
 import 'dart:io';
-import 'package:gpth/date_extractors/json_extractor.dart';
+import 'package:gpth/domain/services/date_extraction/json_date_extractor.dart';
 import 'package:path/path.dart' as p;
 import 'package:test/test.dart';
 import 'test_setup.dart';